requirements.txt
pathlib==1.0.1
pdfplumber==0.11.6
nltk==3.9.1
sentence-transformers==3.4.1
scikit-learn==1.5.2
numpy==1.26.4
mistralai==1.6.0
bs4==0.0.2
requests==2.32.3
faiss-cpu==1.10.0
pathlib==1.0.1